requirements.txt
fastapi
uvicorn
pydantic
motor
fastapi-mail
python-jose
python-dotenv
